bot/requirements.txt
python-telegram-bot==21.7
requests==2.31.0
python-dotenv==1.0.1
